+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.0.16 - 2023-07-02
 ### Added
 - Add command descriptions.
+
 ### Fixed
 - Fix phpstan, psalm and code style problems.
 - Fix pagination problem in GetChildrenController, GetParentsController, GetRelatedController.